move reusable react components to js/forms
