--- src/openzaak/js/components/config/external-services/external-formset.js
@@ -1,10 +1,0 @@
-import React from "react";
-
-
-function ExternalFormSet(props) {
-    return (
-        <p>{props.formData + ''}</p>
-    );
-}
-
-export { ExternalFormSet };
