+++ src/openzaak/js/index.js
@@ -1,2 +1,2 @@
-import './components/config/external-services';
+import './components/external-services';
 import './components/admin';
